refactor: move SaveConfig function to loader to be able to save config after startup
