+++ src/web/login.go
@@ -10,7 +10,6 @@ import (
 	"time"
 
 	"github.com/JacksonTheMaster/StationeersServerUI/v5/src/config"
-	"github.com/JacksonTheMaster/StationeersServerUI/v5/src/configchanger"
 	"github.com/JacksonTheMaster/StationeersServerUI/v5/src/loader"
 	"github.com/JacksonTheMaster/StationeersServerUI/v5/src/logger"
 	"github.com/JacksonTheMaster/StationeersServerUI/v5/src/security"
@@ -192,7 +191,7 @@ func RegisterUserHandler(w http.ResponseWriter, r *http.Request) {
 	existingConfig.Users[creds.Username] = hashedPassword
 
 	// Persist the updated config
-	if err := configchanger.SaveConfig(existingConfig); err != nil {
+	if err := loader.SaveConfig(existingConfig); err != nil {
 		w.Header().Set("Content-Type", "application/json")
 		w.WriteHeader(http.StatusInternalServerError)
 		json.NewEncoder(w).Encode(map[string]string{"error": "Internal Server Error - Failed to save config"})
@@ -235,7 +234,7 @@ func SetupFinalizeHandler(w http.ResponseWriter, r *http.Request) {
 	newConfig.AuthEnabled = &isTrue // Set the pointer to true
 
 	// Save the updated config
-	err = configchanger.SaveConfig(newConfig)
+	err = loader.SaveConfig(newConfig)
 	if err != nil {
 		w.Header().Set("Content-Type", "application/json")
 		w.WriteHeader(http.StatusInternalServerError)
